@@ -30,8 +30,8 @@
     ]
   },
   "dependencies": {
-    "@azure/msal-browser": "^2.24.0",
-    "@azure/msal-react": "^1.4.0",
+    "@azure/msal-browser": "^3.28.0",
+    "@azure/msal-react": "^2.2.0",
     "@mapbox/mapbox-gl-draw": "^1.4.3",
     "@mapbox/mapbox-gl-geocoder": "^5.0.0",
     "@mdi/js": "^6.5.95",
@@ -39,7 +39,6 @@
     "@react-hook/debounce": "^4.0.0",
     "axios": "^1.7.7",
     "bootstrap": "^5.1.3",
-    "bootstrap-switch-button-react": "^1.2.0",
     "canvg": "^4.0.0",
     "clarity-js": "^0.7.10",
     "fast-deep-equal": "^3.1.3",
@@ -49,19 +48,19 @@
     "mapbox-gl": "^3.8.0",
     "mdi-react": "^8.4.0",
     "moment": "^2.29.1",
-    "react": "^17.0.2",
+    "react": "^19.0.0",
     "react-bootstrap": "^2.1.2",
     "react-compound-slider": "^3.3.1",
-    "react-dnd": "^15.1.1",
-    "react-dnd-touch-backend": "^15.1.1",
-    "react-dom": "^17.0.2",
+    "react-dnd": "^16.0.1",
+    "react-dnd-touch-backend": "^16.0.1",
+    "react-dom": "^19.0.0",
     "react-ga4": "^1.4.1",
     "react-multiselect-dropdown-bootstrap": "^1.1.0",
-    "react-query": "^3.34.15",
+    "react-query": "^3.39.3",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.2.2",
-    "react-toastify": "^8.2.0",
+    "react-toastify": "^11.0.0",
     "sass": "^1.80.3",
     "shapefile": "^0.6.6",
     "shpjs": "^6.1.0",
@@ -72,13 +71,18 @@
     "worker-loader": "^3.0.8"
   },
   "overrides": {
-    "typescript": "^5.6.3"
+    "typescript": "^5.6.3",
+    "@azure/msal-react": {
+      "react": "^16.8.0 || ^17 || ^18 || ^19"
+    },
+    "react-query": {
+      "react": "^17 || ^18 || ^19"
+    }
   },
   "devDependencies": {
     "@eslint/js": "^9.12.0",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^13.5.0",
     "@types/chroma-js": "^2.1.3",
     "@types/file-saver": "^2.0.5",
@@ -87,8 +91,8 @@
     "@types/mapbox__mapbox-gl-geocoder": "^4.7.0",
     "@types/mapbox__mapbox-gl-draw": "^1.4.8",
     "@types/node": "^22.7.8",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
     "@types/shapefile": "^0.6.4",
     "@types/shpjs": "^3.4.7",
     "@types/uuid": "^10.0.0",
