@@ -22,6 +22,8 @@ import { FeatureCollection, Feature, GeoJsonProperties, Geometry } from 'geojson
 import { useHomePageContext } from '../home-page/Provider';
 
 import './map.scss';
+import { createRoot } from 'react-dom/client';
+import { ToastContainer } from 'react-toastify';
 interface mapProps {
   handleMapDrawnPolygonChange?: (polygons: Feature<Geometry, GeoJsonProperties>[]) => void;
   handleMapFitChange?: () => void;
@@ -286,7 +288,9 @@ function Map({ handleMapDrawnPolygonChange, handleMapFitChange }: mapProps) {
         })
         .setHTML("<div id='mapboxPopupId'></div>")
         .once('open', () => {
-          ReactDOM.render(mapPopup.element, document.getElementById('mapboxPopupId'));
+          const popupContainer = document.getElementById('mapboxPopupId');
+          const root = createRoot(popupContainer!);
+          root.render(mapPopup.element);
         })
         .addTo(map);
     } else {
@@ -480,7 +484,14 @@ function Map({ handleMapDrawnPolygonChange, handleMapFitChange }: mapProps) {
       )}
       {legend && map && <div className={`legend ${legendClass}`}>{legend}</div>}
       {map && mapAlert}
-      <div id="map" className="map h-100" ref={dropRef}></div>
+      <div
+        id="map"
+        className="map h-100"
+        ref={(el) => {
+          dropRef(el);
+        }}
+      ></div>
+      <ToastContainer />
     </div>
   );
 }
