@@ -82,7 +82,7 @@ function App({ msalInstance }: AppProps) {
               </Route>
             </Routes>
             <ReactQueryDevtools initialIsOpen={false} />
-            <ToastContainer />
+            <ToastContainer containerId="app-toast-container" />
           </DndProvider>
         </QueryClientProvider>
       </AppProvider>
