@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { JSX } from 'react';
 import { mdiInformationOutline } from '@mdi/js';
 import Icon from '@mdi/react';
 import { useState } from 'react';
