@@ -46,7 +46,10 @@ const defaultState: WaterRightDetailsPageContextState = {
 const WaterRightDetailsContext = createContext<WaterRightDetailsPageContextState>(defaultState);
 export const useWaterRightDetailsContext = () => useContext(WaterRightDetailsContext);
 
-export const WaterRightDetailsProvider: FC = ({ children }) => {
+interface WaterRightDetailsProviderProps {
+  children: React.ReactNode;
+}
+export const WaterRightDetailsProvider = ({ children }: WaterRightDetailsProviderProps) => {
   const { id: allocationUuid } = useParams();
 
   const [activeTab, setActiveTab] = useState<ActiveTabType>(defaultState.activeTab);
