@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { JSX } from 'react';
 import { useEffect, useMemo } from 'react';
 import { useMapContext } from '../../../../contexts/MapProvider';
 import { useWaterRightDetailsContext } from '../Provider';
