@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { JSX } from 'react';
 function MainPanel(props: { children: JSX.Element | JSX.Element[] }) {
   return <div className="flex-grow-1 position-relative">{props.children}</div>;
 }
