@@ -1,9 +1,8 @@
 import React from 'react';
 import { useCallback, useMemo, useState } from 'react';
 import Button from 'react-bootstrap/Button';
-import { Accordion } from 'react-bootstrap';
+import { Accordion, Form } from 'react-bootstrap';
 import BeneficialUseSelect from './filters/BeneficialUseSelect';
-import BootstrapSwitchButton from 'bootstrap-switch-button-react';
 import { useWaterRightsContext } from './Provider';
 import { AccordionEventKey } from 'react-bootstrap/esm/AccordionContext';
 import {
@@ -94,10 +93,14 @@ function SideBar() {
           </Accordion.Item>
           <Accordion.Item eventKey="overlayFilters">
             <Accordion.Header onClick={toggleOverlayFilter}>
+              <Form.Check
+                type="switch"
+                id="overlayFilters"
+                label=""
+                checked={isOverlayFilterActive}
+                onChange={toggleOverlayFilter}
+              />
               <label className="fw-bold">OVERLAY FILTER {isOverlayFilterActive}</label>
-              <div className="px-1">
-                <BootstrapSwitchButton checked={isOverlayFilterActive} onstyle="primary" offstyle="secondary" />
-              </div>
             </Accordion.Header>
             <Accordion.Body>
               <Overlays />
@@ -106,10 +109,14 @@ function SideBar() {
 
           <Accordion.Item eventKey="siteSelectionFilters">
             <Accordion.Header onClick={toggleWaterRightFilters}>
+              <Form.Check
+                type="switch"
+                id="waterRightSelection"
+                label=""
+                checked={isWaterRightsFilterActive}
+                onChange={toggleWaterRightFilters}
+              />
               <label className="fw-bold">WATER RIGHT SELECTION {isWaterRightsFilterActive}</label>
-              <div className="px-1">
-                <BootstrapSwitchButton checked={isWaterRightsFilterActive} onstyle="primary" offstyle="secondary" />
-              </div>
             </Accordion.Header>
             <Accordion.Body>
               <div className="mb-3">
@@ -138,10 +145,14 @@ function SideBar() {
           </Accordion.Item>
           <Accordion.Item eventKey="nldi">
             <Accordion.Header onClick={toggleNldiFilterStatus}>
+              <Form.Check
+                type="switch"
+                id="nldi"
+                label=""
+                checked={isNldiFilterActive}
+                onChange={toggleNldiFilterStatus}
+              />
               <label className="fw-bold">NLDI FILTER {isNldiFilterActive}</label>
-              <div className="px-5">
-                <BootstrapSwitchButton checked={isNldiFilterActive} onstyle="primary" offstyle="secondary" />
-              </div>
             </Accordion.Header>
             <div className="px-4" style={{ background: 'ghostwhite' }}>
               <a
