@@ -1,4 +1,4 @@
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook } from '@testing-library/react';
 import { useNldiFilter } from './useNldiFilter';
 import * as WaterRightsProvider from '../../Provider';
 import * as NldiQueries from '../../../../../hooks/queries/useNldiQuery';
